Issue #3399992 by mstrelan, smustgrave, xjm: Fix strict type errors in test traits
(cherry picked from commit c304b09272b6e760db8cba83e718df454e20df7a)

+++ core/modules/filter/tests/src/Kernel/FilterKernelTest.php
@@ -391,7 +391,7 @@ class FilterKernelTest extends KernelTestBase {
     $this->assertFilteredString($filter, $tests);
 
     // Very long string hitting PCRE limits.
-    $limit = max(ini_get('pcre.backtrack_limit'), ini_get('pcre.recursion_limit'));
+    $limit = max((int) ini_get('pcre.backtrack_limit'), (int) ini_get('pcre.recursion_limit'));
     $source = $this->randomMachineName($limit);
     $result = _filter_autop($source);
     $this->assertEquals($result, '<p>' . $source . "</p>\n", 'Line break filter can process very long strings.');
